+++ src/promptflow/promptflow/_sdk/operations/_local_azure_connection_operations.py
@@ -9,7 +9,7 @@ from promptflow._sdk._telemetry import ActivityType, WorkspaceTelemetryMixin, mo
 from promptflow._sdk._utils import interactive_credential_disabled, is_from_cli, is_github_codespaces, print_red_error
 from promptflow._sdk.entities._connection import _Connection
 from promptflow._utils.logger_utils import get_cli_sdk_logger
-from promptflow.azure._utils.gerneral import get_arm_token
+from promptflow.azure._utils.general import get_arm_token
 
 logger = get_cli_sdk_logger()
