+++ src/promptflow/promptflow/azure/_restclient/flow_service_caller.py
@@ -19,7 +19,7 @@ from promptflow._sdk._telemetry import TelemetryMixin
 from promptflow._utils.logger_utils import LoggerFactory
 from promptflow.azure._constants._flow import AUTOMATIC_RUNTIME, SESSION_CREATION_TIMEOUT_ENV_VAR
 from promptflow.azure._restclient.flow import AzureMachineLearningDesignerServiceClient
-from promptflow.azure._utils.gerneral import get_authorization, get_arm_token, get_aml_token
+from promptflow.azure._utils.general import get_authorization, get_arm_token, get_aml_token
 from promptflow.exceptions import UserErrorException, PromptflowException, SystemErrorException
 
 logger = LoggerFactory.get_logger(__name__)
