+++ src/promptflow/promptflow/azure/_utils/__init__.py
@@ -4,6 +4,6 @@
 
 __path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore
 
-from .gerneral import is_arm_id
+from .general import is_arm_id
 
 __all__ = ["is_arm_id"]
