+++ src/promptflow/promptflow/azure/_utils/general.py
@@ -3,6 +3,7 @@
 # ---------------------------------------------------------
 
 import jwt
+
 from promptflow.exceptions import ValidationException
